# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trackvance Core",
    "description": "Monolito modular local. SUCCESS es estado técnico; decision conserva el resultado de negocio. Configuraciones y evidence son snapshots versionados.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Usuarios locales"
        ],
        "summary": "Users",
        "operationId": "users_api_v1_users_get",
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "users:read"
      },
      "post": {
        "tags": [
          "Usuarios locales"
        ],
        "summary": "Create User",
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "users:write",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/users/roles": {
      "get": {
        "tags": [
          "Usuarios locales"
        ],
        "summary": "Roles",
        "operationId": "roles_api_v1_users_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "users:read"
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "Usuarios locales"
        ],
        "summary": "User Detail",
        "operationId": "user_detail_api_v1_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "users:read"
      },
      "patch": {
        "tags": [
          "Usuarios locales"
        ],
        "summary": "Edit User",
        "operationId": "edit_user_api_v1_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "users:write"
      }
    },
    "/api/v1/users/{user_id}/reset-password": {
      "post": {
        "tags": [
          "Usuarios locales"
        ],
        "summary": "Reset Password",
        "operationId": "reset_password_api_v1_users__user_id__reset_password_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "users:write"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordReset": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "password"
        ],
        "title": "PasswordReset"
      },
...
      "UserCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Email"
          },
          "role": {
            "type": "string",
            "enum": [
              "Administrator",
              "Data Owner",
              "Data Owner / Lead",
              "Data Analyst",
              "Operations",
              "Auditor"
            ],
            "title": "Role",
            "default": "Data Analyst"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "UserListResponse"
      },
      "UserPatch": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Administrator",
                  "Data Owner",
                  "Data Owner / Lead",
                  "Data Analyst",
                  "Operations",
                  "Auditor"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version"
        ],
        "title": "UserPatch"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "password_changed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Changed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "active",
          "permissions",
          "version",
          "created_at",
          "updated_at",
          "password_changed_at"
        ],
        "title": "UserResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "LocalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "trackvance_session",
        "description": "Sesión local HttpOnly. Las mutaciones requieren además X-CSRF-Token."
      }
    }
  }
}